[JENKINS-57138] Optimize filter loop + externalize WordTrie
